listView style
made a bewatiful listview

+++ app/src/main/java/com/example/android/getmoviex/PixelCalc.java
@@ -8,7 +8,7 @@ import android.util.TypedValue;
  */
 
 public class PixelCalc {
-    public static int getPixels(float dp){
+    public static int getPixels(int dp){
         DisplayMetrics displayMetrics=MyApp.getContext().getResources().getDisplayMetrics();
         float px= TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP,dp,displayMetrics);
         return Math.round(px);
